+++ src/app/(after)/recommend/page.tsx
@@ -18,7 +18,7 @@ import NavigationBar from '@/components/NavigationBar/NavigationBar'
 import PlansModal from '@/components/Modal/PlansModal'
 import { useModal } from '@/hooks/useModal/useModal'
 import { useTranslate } from '@/hooks/useTranslate/useTranslate'
-import { useWeatherContext } from '@/providers/WeatherProviter'
+import { useWeatherContext } from '@/providers/WeatherProvider'
 import { RootState } from '@/redux/store'
 import { useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
